Routing portfolio search Phase 2a: extract shared routing scorer
Move the metric computation out of the harness (routingHarness/metrics.ts) into
src/routing/scoreRoutes.ts (computeRoutingMetrics + scoreRoutes). The harness's
computeRuleReport now wraps it. This gives ONE definition of routing quality
shared by the offline harness gate and the in-worker portfolio scorer (Phase 2b)
— so the search optimizes exactly the objective the harness validates, not a
drifting copy. Pure code move; routing:check byte-identical.

tsc -b clean, vitest 178/178, routing:check unchanged.
